+++ hermes-management/src/main/java/pl/allegro/tech/hermes/management/config/SchemaRepositoryProperties.java
@@ -19,6 +19,8 @@ public class SchemaRepositoryProperties {
 
     private boolean subjectNamespaceEnabled = Configs.SCHEMA_REPOSITORY_SUBJECT_NAMESPACE_ENABLED.getDefaultValue();
 
+    private SchemaRepositoryAuthorizationProperties authorization = new SchemaRepositoryAuthorizationProperties();
+
     public String getServerUrl() {
         return serverUrl;
     }
@@ -73,5 +75,13 @@ public class SchemaRepositoryProperties {
 
     public void setSubjectNamespaceEnabled(boolean subjectNamespaceEnabled) {
         this.subjectNamespaceEnabled = subjectNamespaceEnabled;
+    }
+
+    public SchemaRepositoryAuthorizationProperties getAuthorization() {
+        return authorization;
+    }
+
+    public void setAuthorization(SchemaRepositoryAuthorizationProperties authorization) {
+        this.authorization = authorization;
     }
 }